fix opengraph
