+++ app/(main)/[lang]/data.tsx
@@ -34,11 +34,6 @@ import {
   mainFaq3Answer
 } from "@/locales/.generated/server";
 
-export const data = {
-  name: "Hyperjump Technology",
-  baseUrl: "https://hyperjump.tech"
-};
-
 export const mainNav = (lang: SupportedLanguage) => {
   const data = [
     { label: mainNavItems0Label(lang), href: "/services" },
